img_upd4: Table adjust and Unix-like software
