+++ lang/en-US/3bcce27b-3a3a-4a20-b294-1b98f7f72728.json
@@ -30,7 +30,7 @@
       "CC BY-NC 4.0 DEED"
    ],
    "last_update": [
-      "1771028015"
+      "1780448234"
    ],
    "authors": [
       ""
@@ -306,7 +306,7 @@
                "PostMention": ""
             },
             {
-               "text": "<style>.localtablename {.width: 60%;margin-left: auto;margin-right: auto;text-align: center;border-collapse: collapse;} .localtablename th {border-collapse: collapse;background-color:lightgrey;} .localtablename td:nth-child(-n+1) {background-color:lightblue;}</style><p><table class=\"localtablename\"><tr><th>Number</th><th>Family</th><th>Prefix</th></tr> <tr><td>1,000,000,000,000,000,000</td><td>Quintillion</td><td>Hexa</td> </tr> <tr><td>1,000,000,000,000,000</td><td>Quadrillion</td><td>Peta</td> </tr>  <tr><td>1,000,000,000,000</td><td>Trillion</td><td>Tera</td> </tr> <tr><td>1,000,000,000</td><td>Billion</td><td>Giga</td> </tr> <tr><td>1,000,000</td><td>Million</td><td>Mega</td> </tr> <tr><td>1,000</td><td>Thousand</td><td>Kilo</td> </tr><tr><td>1</td><td>Units</td><td>&nbsp;</td> </tr> <tr><td>0.1</td><td>Tenths</td><td>Deci</td> </tr> <tr><td>0.01</td><td>Hundredths</td><td>Centi</td> </tr> <tr><td>0.001</td><td>Thousandths</td><td>Mili</td> </tr> <tr><td>0.000001</td><td>Millionth</td><td>Micro</td> </tr> <tr><td>0.000000001</td><td>Billionth</td><td>Nano</td> </tr> <tr><td>0.000000000001</td><td>Trillionth</td><td>Pico</td> </tr> </table></p><p class=\"desc\"><b>Table 1</b>: Names assigned to numbers.</p>",
+               "text": "<style>.localtablename {width: 60%;display: block;overflow-x: auto;max-width: 100%;margin-left: auto;margin-right: auto;text-align: center;border-collapse: collapse;} .localtablename th {border-collapse: collapse;background-color:lightgrey;} .localtablename td:nth-child(-n+1) {background-color:lightblue;}</style><p><table class=\"localtablename\"><tr><th>Number</th><th>Family</th><th>Prefix</th></tr> <tr><td>1,000,000,000,000,000,000</td><td>Quintillion</td><td>Hexa</td> </tr> <tr><td>1,000,000,000,000,000</td><td>Quadrillion</td><td>Peta</td> </tr>  <tr><td>1,000,000,000,000</td><td>Trillion</td><td>Tera</td> </tr> <tr><td>1,000,000,000</td><td>Billion</td><td>Giga</td> </tr> <tr><td>1,000,000</td><td>Million</td><td>Mega</td> </tr> <tr><td>1,000</td><td>Thousand</td><td>Kilo</td> </tr><tr><td>1</td><td>Units</td><td>&nbsp;</td> </tr> <tr><td>0.1</td><td>Tenths</td><td>Deci</td> </tr> <tr><td>0.01</td><td>Hundredths</td><td>Centi</td> </tr> <tr><td>0.001</td><td>Thousandths</td><td>Mili</td> </tr> <tr><td>0.000001</td><td>Millionth</td><td>Micro</td> </tr> <tr><td>0.000000001</td><td>Billionth</td><td>Nano</td> </tr> <tr><td>0.000000000001</td><td>Trillionth</td><td>Pico</td> </tr> </table></p><p class=\"desc\"><b>Table 1</b>: Names assigned to numbers.</p>",
                "source": null,
                "date_time": null,
                "isTable": false,
