+++ lang/es-ES/33c83383-b9b5-43ec-987d-68f027f342b2.json
@@ -29,7 +29,7 @@
       "CC BY-NC 4.0 DEED"
    ],
    "last_update": [
-      "1780408032"
+      "1780448198"
    ],
    "authors": "",
    "reviewers": "",
@@ -1061,7 +1061,7 @@
                "PostMention": "."
             },
             {
-               "text": "<style>.localtable {.width: 100%;margin-left: auto;margin-right: auto;text-align: center; border-collapse: separate; border-spacing: 2px;} .localtable th {border-collapse: collapse;background-color:lightgrey; border: 1px solid black;} .localtable td:nth-child(-n+1) {background-color:lightblue;width: 36.36%;} .localtable td {background-color:white;width: 6.36%; border-collapse: collapse; border: 1px solid black;}</style><p><table class=\"localtable\"><tr><th>Característica</th><th><i>Homo sapiens</i></th><th><i>Homo rudolfensis</i></th><th><i>Homo habilis</i></th><th><i>Homo erectus</i></th><th><i>Homo ergaster</th><th><i>Homo heidelbergensis</i></th><th><i>Homo naledi</i></th><th><i>Homo neanderthalensis</i></th><th><i>Homo floresiensis</i></th><th><i>Homo sapiens</i></th></tr> <tr><td>Dos ojos</td><td><i>X</i></td><td><i>X</i></td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td></tr> <tr><td>Gran volumen cerebral (más de 1300 ml)</td><td><i>X</i></td><td><i>&nbsp;</i></td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td></tr> <tr><td>&nbsp</td><td><i>&nbsp;</i></td><td><i>&nbsp;</i></td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td></tr></table></p><p class=\"desc\"><b>Tabla 1</b>: Ejemplo de tabla de comparación.</p>",
+               "text": "<style>.localtable {width: 100%;display: block;overflow-x: auto;max-width: 100%;margin-left: auto;margin-right: auto;text-align: center; border-collapse: separate; border-spacing: 2px;} .localtable th {border-collapse: collapse;background-color:lightgrey; border: 1px solid black;} .localtable td:nth-child(-n+1) {background-color:lightblue;width: 36.36%;} .localtable td {background-color:white;width: 6.36%; border-collapse: collapse; border: 1px solid black;}</style><p><table class=\"localtable\"><tr><th>Característica</th><th><i>Homo sapiens</i></th><th><i>Homo rudolfensis</i></th><th><i>Homo habilis</i></th><th><i>Homo erectus</i></th><th><i>Homo ergaster</th><th><i>Homo heidelbergensis</i></th><th><i>Homo naledi</i></th><th><i>Homo neanderthalensis</i></th><th><i>Homo floresiensis</i></th><th><i>Homo sapiens</i></th></tr> <tr><td>Dos ojos</td><td><i>X</i></td><td><i>X</i></td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td></tr> <tr><td>Gran volumen cerebral (más de 1300 ml)</td><td><i>X</i></td><td><i>&nbsp;</i></td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td></tr> <tr><td>&nbsp</td><td><i>&nbsp;</i></td><td><i>&nbsp;</i></td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td></tr></table></p><p class=\"desc\"><b>Tabla 1</b>: Ejemplo de tabla de comparación.</p>",
                "source": null,
                "date_time": null,
                "isTable": false,
