+++ lang/pt-BR/3bcce27b-3a3a-4a20-b294-1b98f7f72728.json
@@ -30,7 +30,7 @@
       "CC BY-NC 4.0 DEED"
    ],
    "last_update": [
-      "1771028025"
+      "1780448263"
    ],
    "authors": [
       ""
@@ -306,7 +306,7 @@
                "PostMention": ""
             },
             {
-               "text": "<style>.localtablename {.width: 60%;margin-left: auto;margin-right: auto;text-align: center;border-collapse: collapse;} .localtablename th {border-collapse: collapse;background-color:lightgrey;} .localtablename td:nth-child(-n+1) {background-color:lightblue;}</style><p><table class=\"localtablename\"><tr><th>Número</th><th>Família</th><th>Prefixo</th></tr> <tr><td>1.000.000.000.000.000.000</td><td>Quintrilhão</td><td>Hexa</td> </tr> <tr><td>1.000.000.000.000.000</td><td>Quatrilhão</td><td>Peta</td> </tr>  <tr><td>1.000.000.000.000</td><td>Trilhão</td><td>Tera</td> </tr> <tr><td>1.000.000.000</td><td>Bilhão</td><td>Giga</td> </tr> <tr><td>1.000.000</td><td>Milhão</td><td>Mega</td> </tr> <tr><td>1.000</td><td>Mil</td><td>Quilo</td> </tr><tr><td>1</td><td>Unidades</td><td>&nbsp;</td> </tr> <tr><td>0,1</td><td>Décimos</td><td>Deci</td> </tr> <tr><td>0,01</td><td>Centésimos</td><td>Centi</td> </tr> <tr><td>0,001</td><td>Milésimos</td><td>Mili</td> </tr> <tr><td>0,000001</td><td>Milionésimo</td><td>Micro</td> </tr> <tr><td>0,000000001</td><td>Bilionésimo</td><td>Nano</td> </tr> <tr><td>0,000000000001</td><td>Trilhonésimo</td><td>Pico</td> </tr> </table></p><p class=\"desc\"><b>Tabela 1</b>: Nomes atribuídos aos números.</p>",
+               "text": "<style>.localtablename {width: 60%;display: block;overflow-x: auto;max-width: 100%;margin-left: auto;margin-right: auto;text-align: center;border-collapse: collapse;} .localtablename th {border-collapse: collapse;background-color:lightgrey;} .localtablename td:nth-child(-n+1) {background-color:lightblue;}</style><p><table class=\"localtablename\"><tr><th>Número</th><th>Família</th><th>Prefixo</th></tr> <tr><td>1.000.000.000.000.000.000</td><td>Quintrilhão</td><td>Hexa</td> </tr> <tr><td>1.000.000.000.000.000</td><td>Quatrilhão</td><td>Peta</td> </tr>  <tr><td>1.000.000.000.000</td><td>Trilhão</td><td>Tera</td> </tr> <tr><td>1.000.000.000</td><td>Bilhão</td><td>Giga</td> </tr> <tr><td>1.000.000</td><td>Milhão</td><td>Mega</td> </tr> <tr><td>1.000</td><td>Mil</td><td>Quilo</td> </tr><tr><td>1</td><td>Unidades</td><td>&nbsp;</td> </tr> <tr><td>0,1</td><td>Décimos</td><td>Deci</td> </tr> <tr><td>0,01</td><td>Centésimos</td><td>Centi</td> </tr> <tr><td>0,001</td><td>Milésimos</td><td>Mili</td> </tr> <tr><td>0,000001</td><td>Milionésimo</td><td>Micro</td> </tr> <tr><td>0,000000001</td><td>Bilionésimo</td><td>Nano</td> </tr> <tr><td>0,000000000001</td><td>Trilhonésimo</td><td>Pico</td> </tr> </table></p><p class=\"desc\"><b>Tabela 1</b>: Nomes atribuídos aos números.</p>",
                "source": null,
                "date_time": null,
                "isTable": false,
