+++ src/viewer/go.mod
@@ -10,3 +10,5 @@ require (
 	github.com/x448/float16 v0.8.4 // indirect
 	golang.org/x/sys v0.43.0 // indirect
 )
+
+require github.com/webview/webview_go v0.0.0-20240831120633-6173450d4dd6
